+++ Intrepion.ToDo/Controllers/ApplicationUserController.cs
@@ -12,9 +12,14 @@ public class AdminApplicationUserEditDataTransferObjectController(IApplicationUs
     private readonly IApplicationUserService _applicationUserService = applicationUserService;
 
     [HttpPost]
-    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject>> Add(AdminApplicationUserEditDataTransferObject applicationUser)
+    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject?>> Add(AdminApplicationUserEditDataTransferObject applicationUser)
     {
         var userName = User.Identity?.Name;
+
+        if (userName == null)
+        {
+            return Ok(null);
+        }
 
         var addedAdminApplicationUserEditDataTransferObject = await _applicationUserService.AddAsync(userName, applicationUser);
 
@@ -22,9 +27,14 @@ public class AdminApplicationUserEditDataTransferObjectController(IApplicationUs
     }
 
     [HttpDelete("{id}")]
-    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject>> Delete(string id)
+    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject?>> Delete(string id)
     {
         var userName = User.Identity?.Name;
+
+        if (userName == null)
+        {
+            return Ok(null);
+        }
 
         var result = await _applicationUserService.DeleteAsync(userName, id);
 
@@ -32,9 +42,14 @@ public class AdminApplicationUserEditDataTransferObjectController(IApplicationUs
     }
 
     [HttpPut("{id}")]
-    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject>> Edit(string id, AdminApplicationUserEditDataTransferObject applicationUser)
+    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject?>> Edit(string id, AdminApplicationUserEditDataTransferObject applicationUser)
     {
         var userName = User.Identity?.Name;
+
+        if (userName == null)
+        {
+            return Ok(null);
+        }
 
         var updatedAdminApplicationUserEditDataTransferObject = await _applicationUserService.EditAsync(userName, id, applicationUser);
 
@@ -42,16 +57,30 @@ public class AdminApplicationUserEditDataTransferObjectController(IApplicationUs
     }
 
     [HttpGet]
-    public async Task<ActionResult<List<AdminApplicationUserListItemDataTransferObject>>> GetAll()
+    public async Task<ActionResult<List<AdminApplicationUserListItemDataTransferObject>?>> GetAll()
     {
+        var userName = User.Identity?.Name;
+
+        if (userName == null)
+        {
+            return Ok(null);
+        }
+
         var applicationUsers = await _applicationUserService.GetAllAsync();
 
         return Ok(applicationUsers);
     }
 
     [HttpGet("{id}")]
-    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject>> GetById(string id)
+    public async Task<ActionResult<AdminApplicationUserEditDataTransferObject?>> GetById(string id)
     {
+        var userName = User.Identity?.Name;
+
+        if (userName == null)
+        {
+            return Ok(null);
+        }
+
         var applicationUser = await _applicationUserService.GetByIdAsync(id);
 
         return Ok(applicationUser);
